catalog: carve FlowRegistrationService out of the facade
Move flow registration CRUD (register, update, delete, get, list,
list_artifacts), the namespace-seeding helpers (ensure_*,
_ensure_general_child), resolve_registration_id, and bulk_enrich_flows
into catalog/services/flows.py. Promote _bulk_enrich_flows /
_enrich_flow_registration to public on the sub-service for cross-
service consumption.

auto_register_flow stays on the facade — a test monkeypatches
CatalogService.ensure_local_flows_namespace and routing the ensure_*
calls through self preserves that contract.
